requirements.txt
websockets
python-dotenv
tinydb